# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EarScope API",
        "description": "Dokumentasi API untuk EarScope",
        "version": "1.0"
    },
    "paths": {
        "/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "operationId": "7abb047ced1329103f0767160e852fc9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_akses",
                                    "password"
                                ],
                                "properties": {
                                    "kode_akses": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "ERS-736521"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout User",
                "operationId": "fb4913f284432ecf8c1394ad6c99a14e",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh JWT Token",
                "operationId": "5a69b0db2ba666122033a1ee3e9e582e",
                "responses": {
                    "200": {
                        "description": "Token refreshed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token di kolom ini dengan format Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}